refactor(repo.mapper): moved Repo Api types to separate file to reuse in tests
