+++ src/components/search-bar/search-bar.constants.ts
@@ -1,6 +1,5 @@
+import { RequestSortBy } from 'components/repos/repo.mapper.types';
 import { SortBy } from 'components/repos/repos.types';
-
-import { RequestSortBy } from 'components/repos/repo.mapper';
 
 export const sortOptions: { value: SortBy, label: string }[] = [
     { value: 'stars', label: 'Stars' },
