+++ src/scripts/search.script.ts
@@ -1,7 +1,8 @@
 import { makeAutoObservable, runInAction } from 'mobx';
 import { TransactionScript } from 'scripts/transaction-script';
 
-import { repoMapper, RequestSortBy } from 'components/repos/repo.mapper';
+import { repoMapper } from 'components/repos/repo.mapper';
+import { RequestSortBy } from 'components/repos/repo.mapper.types';
 import { Repo } from 'components/repos/repos.types';
 
 export class SearchTS implements TransactionScript {
